Refactor: 여러 부분 리팩토링

+++ src/main/java/nbc/chillguys/nebulazone/application/user/dto/response/UserResponse.java
@@ -18,7 +18,7 @@ public record UserResponse(
 	String phone,
 	String nickname,
 	String profileImageUrl,
-	int point,
+	long point,
 	OAuthType oAuthType,
 	String oAuthId,
 	Set<AddressResponse> addresses,
